+++ packages/components/components/elvis-radio-filter/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Elvia Radio Filter Changelog
 
-## 1.1.1 (23.11.21)
+## 1.1.2 (30.11.21)
 
 ### Bug fix
 
 - Now using elvis-colors package for colors.
+
+## 1.1.1 (30.11.21)
+
+- Web component wrapper fix
 
 ## 1.1.0 (03.11.21)